+++ app/src/main/java/edu/mobile/voterlist/api/DataFileApi.java
@@ -1,11 +1,7 @@
 package edu.mobile.voterlist.api;
 
-import java.io.InputStream;
-
-import edu.mobile.voterlist.InputStreamRequestBody;
 import edu.mobile.voterlist.model.DataFileInfo;
 import okhttp3.MultipartBody;
-import okhttp3.RequestBody;
 import okhttp3.ResponseBody;
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -20,7 +16,7 @@ public interface DataFileApi {
     @POST("/api/uploadFile")
     Call<DataFileInfo> save(@Part MultipartBody.Part file);
     @GET("/api/uploadFile/{id}/content")
-    Call<ResponseBody> getUserProfileImage(@Path("id") int id);
+    Call<ResponseBody> getUserProfileImage(@Path("id") long id);
 
     @GET("/api/uploadFile/{id}")
     Call<DataFileInfo> getUserProfile(@Path("id") int id);
